moved saving street thumbnail functions to streets/image; updated SaveAsImageDialog test
