+++ assets/scripts/dialogs/SaveAsImageDialog.jsx
@@ -10,8 +10,7 @@ import { connect } from 'react-redux'
 import { FormattedMessage, FormattedHTMLMessage, injectIntl, intlShape } from 'react-intl'
 import Dialog from './Dialog'
 import { trackEvent } from '../app/event_tracking'
-import { getStreetImage } from '../streets/image'
-import { saveStreetThumbnail, checkSaveThumbnailIncomplete } from '../streets/thumbnail'
+import { getStreetImage, saveStreetThumbnail, checkSaveThumbnailIncomplete } from '../streets/image'
 import { setSettings } from '../store/actions/settings'
 import { normalizeSlug } from '../util/helpers'
 import './SaveAsImageDialog.scss'
