+++ assets/scripts/dialogs/__tests__/SaveAsImageDialog.test.js
@@ -6,7 +6,9 @@ import { shallowWithIntl as shallow } from '../../../../test/helpers/intl-enzyme
 // Mock dependencies that could break tests
 jest.mock('../../streets/image', () => {
   return {
-    getStreetImage: () => {}
+    getStreetImage: () => {},
+    saveStreetThumbnail: () => {},
+    checkSaveThumbnailIncomplete: () => {}
   }
 })
 jest.mock('../../users/settings', () => {})
